Configuration proceeds smoothly with UseApps
The idea is that people will configure and build the
library, then be able to develop on the apps in ./android
using Eclipse + the generated .mk scripts.

I'm still working out how to handle multiple architectures.
Let's just get arm working so that I can prototype. =D
I want to see this running on my phone, not just an EMU.